Update after the change to type resolution in llvm-hs (#81)
Since llvm-hs now correctly implements type lookup through NamedTypeReferences, `typeOf` is an operation in `MonadModuleBuilder`, requiring lots of small changes in llvm-hs-pretty. The biggest user-facing change is just that the `ppll` operation is now only available on the subset of LLVM IR entities which can be printed without doing a type lookup.

For all other entities, there is now an explicit operation ppEntity in `MonadModuleBuilder`, for example, `ppOperand :: MonadModuleBuilder m => Operand -> m (Doc ann)`. In my personal experience using llvm-hs to build a significant project, the additional `MonadModuleBuilder` constraint is not terribly onerous. Pretty printing snippets of IR is usually something I want to do in the context of constructing a module, to see what I'm generating in a nice format.

Documentation and tests updated.